Fix Streamlit page link by moving panel script
